test(harness): one driveUntil, in the framework
Suite 96 needed the paced chain drive that suite 55 already had, and grew
its own. Two copies of "how fast may a suite drive blocks" is two places
for the answer to drift, and the answer is the one that cost a gate: at
200ms one suite drove about a thousand blocks per wait, held a runner
slot for the full 1800s wall clock, and starved two unrelated suites into
the same wall.

It moves to `daemon-control.js`, which already owns `advanceBlock` and
the ticker, and takes A CLIENT rather than a node index - the two copies
disagreed about whether the index was 0- or 1-based, which is not a
mistake worth leaving available.

`timeoutMs` is now required rather than defaulted. Every deadline here is
wall-clock and each caller derives its own from something different - a
departure cycle, an election cycle - so a shared default is one suite's
number silently applied to another suite's wait. Seven of suite 55's
eight call sites already passed it; the eighth now does, with the same
value it was getting from the default.

`blockIntervalMs` keeps its old default by the same expression suite 55
used for it, so the pacing is unchanged. The error message is unchanged
when no label is given, and takes a `label:` prefix for callers that want
to say what they were waiting for.
